Add Switch component (incomplete)
Switch component is added, but it still filled with bugs.

Switch.png image is a placeholder.

Fix bugs in MockSwitch (still buggy)

Debug Switch component on phone

Toggle switch component now appears to work on the phone with four colors
configurable and events working.

MockSwitch is still a placeholder Checkbox and is under development.

Remove non-working code

Functions on phone, but resizing in designer is a problem.

Change-Id: I2e8801d48b95f58528a40bc15ee4dc3c19ba1bb4